rename the package and classes  to lower case
