+++ lib/ui/main/main.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:sellio_mobile/core/design_system/themes/sellio_theme_provider.dart';
-import 'package:sellio_mobile/ui/screens/auth/login.dart';
+import '../screens/main/main_screen.dart';
 
 
 void main() {
@@ -20,7 +20,7 @@ class MyApp extends StatelessWidget {
     return MaterialApp(
       debugShowCheckedModeBanner: false,
       title: 'Sellio app',
-      home: const LoginScreen(),
+      home: const MainScreen(),
     );
   }
 }
